wrap help output in details block...
Since it's my job to be pedantic... 😄

Just help is an argument, and --help is a flag. You've described it here as the former, but the help output itself describes it as the latter. It looks like currently kc doesn't care and takes either, or it treats one the same as no valid argument, which provides the same output.

As we define this tool more and create more error handling and docs-from-code, it might behoove us to be super clear here on the distinction between arguments and flags, and which (or both) are used for the help output.